requirements.txt
tensorflow>=2.16.0
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.26.0
PyYAML>=6.0
tqdm>=4.66.0
pytest>=7.4.0
